requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1
alembic>=1.11.0
jinja2>=3.1.2
aiofiles>=23.1.0
langchain>=0.0.267
langchain-openai>=0.0.2
chromadb>=0.4.13
openai>=1.2.3
python-dotenv>=1.0.0
pytest>=7.3.1
httpx>=0.24.0
tenacity>=8.2.2
colorlog>=6.7.0
pydantic-settings>=2.0.0
PyPDF2>=3.0.1
playwright>=1.40.0
python-docx>=0.8.11
pandas>=2.0.0
mammoth>=1.6.0
beautifulsoup4>=4.12.0